+++ tests/testthat/test_proportional_hazards.R
@@ -9,12 +9,3 @@ test_that("updating", {
 test_that("bad input", {
   expect_error(proportional_hazards(mode = ", classification"))
 })
-
-test_that("wrong fit interface", {
-  expect_error(
-    expect_message(
-      proportional_hazards() %>% fit_xy()
-    ),
-    "must use the formula interface"
-  )
-})